refactor: dataset_fetcher
* Renames dataset_fetcher to dataframe_fetcher

* Turns DataframeFetcher into an abstract class

* Turns fetch() into an abstract method, and handles the impl
  on Implementation classes

* Changes dataframe col names inflection to snake_case
