+++ week_1_basics_n_setup/pandas_sqlalchemy/pyproject.toml
@@ -19,6 +19,7 @@ dependencies = [
     "psycopg-binary>=3.1.13,<4.0",
     "mysql-connector-python>=8.2.0,<9.0",
     "omegaconf>=2.3.0,<3.0",
+    "inflection>=0.5.1,<1.0",
 ]
 
 [tool.pdm.dev-dependencies]
